# swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "CurificBackend Copy",
		"contact": {},
		"version": "1.0"
	},
	"servers": [
		{
			"url": "http://65.2.142.187:5001/api/v1",
			"variables": {}
		}
	],
	"paths": {
...
		"/user/customer/delete/{id}": {
			"delete": {
				"tags": ["Dr &customer"],
				"summary": "DeleteCustomer",
				"description": "Delete Customer",
				"operationId": "DeleteCustomer",
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m2002"
								},
								"example": {
									"success": true,
									"msg": "Doctor delete successfully",
									"data": []
								}
							}
						}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
		"/user/social/login": {
			"post": {
				"tags": ["Dr &customer"],
				"summary": "sociallogin",
				"description": "social login",
				"operationId": "sociallogin",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/socialloginrequest"
							},
							"example": {
								"token": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6ImNkNjcyYTI5OGRkNTk3NzYyMTJlNmE2ZDgzZmI4YTllNjBmZjNiZWYifQ.eyJpc3MiOiJodHRwczpcL1wvd3d3LmZhY2Vib29rLmNvbSIsImF1ZCI6IjEwOTYwMTA0MzQzMzgxOTIiLCJzdWIiOiIxNDUxMTA5MTg1MzAxNTU0IiwiaWF0IjoxNjU4MzIyMTI3LCJleHAiOjE2NTgzMjU3MjcsImp0aSI6IkRJMFguZDlmMWU3ODZjOTc5OTNkNTk1ODMxZTc1ZTAyNzc0YmNmZDQyODVjMDE5NzFkOGQ4M2NmNWE1M2UzNTJkZTEzMiIsIm5vbmNlIjoibXlfbm9uY2UiLCJlbWFpbCI6InBpcGFsaXlhc2Fua2V0MVx1MDA0MGdtYWlsLmNvbSIsImdpdmVuX25hbWUiOiJTYW5rZXQiLCJmYW1pbHlfbmFtZSI6IlBpcGFsaXlhIiwibmFtZSI6IlNhbmtldCBQaXBhbGl5YSIsInBpY3R1cmUiOiJodHRwczpcL1wvcGxhdGZvcm0tbG9va2FzaWRlLmZic2J4LmNvbVwvcGxhdGZvcm1cL3Byb2ZpbGVwaWNcLz9hc2lkPTE0NTExMDkxODUzMDE1NTQmaGVpZ2h0PTEwMCZ3aWR0aD0xMDAmZXh0PTE2NjA5MTQxMjgmaGFzaD1BZVNVUmR5QVYzX3NaLVkwRDBVIn0.fzXmN3ivyVpq3cWjPV94lgk3iimF6dsl4PnhfvGj2WntwXG8VUJ6o_VEM_Tau_Un6Un_JUl1TnKMM-4ljtCNlEn4YXfGpSHXwatZ22X7UyRjxTTBZezloUw_sFrolSWz_ys9NsKWfGvO8WBLbekppE6bJNUzVe-jCrkLmY7fOVc2IguVg36VWcyIsZAvdeGzsIC5rm3D9jRg8h2HJuTpH5NB2X58Vq8enMv9ZzwGKmRkkFmpUDEyMXGMUZ6eH622-Ngn2hfHIQJIaJFMLEVNxpQMD7GGKNyKSRpuqHK-E-T95Ug-KG1wha6hwi-G_t8So7T2Vqpl8XOCcpJMVRd_Jw",
								"type": "CUSTOMER"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false
			}
		},
...
		"/device/update/{id}": {
			"put": {
				"tags": ["Device"],
				"summary": "Updatedevice",
				"description": "Update device",
				"operationId": "Updatedevice",
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatedeviceRequest"
							},
							"example": {
								"deviceName": "fddddff",
								"deviceType": "aaa",
								"deviceModal": "tm005a",
								"treatment": "aaaa"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m2007"
								},
								"example": {
									"success": true,
									"message": "Device update successfully",
									"data": []
								}
							}
						}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
		"/device/delete/{id}": {
			"delete": {
				"tags": ["Device"],
				"summary": "DeleteDevice",
				"description": "Delete Device",
				"operationId": "DeleteDevice",
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m2007"
								},
								"example": {
									"success": true,
									"message": "Device delete successfully",
									"data": []
								}
							}
						}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
		"/dashboard": {
			"get": {
				"tags": ["Dashboard"],
				"summary": "GetDashboard",
				"description": "Get Dashboard",
				"operationId": "GetDashboard",
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
...
		"/faq/delete/{id}": {
			"delete": {
				"tags": ["faqs"],
				"summary": "DeleteFaqs",
				"description": "Delete Faqs",
				"operationId": "DeleteFaqs",
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"m2002": {
				"title": "m2002",
				"required": ["success", "msg", "data"],
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"msg": {
						"type": "string"
					},
					"data": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"description": ""
					}
				},
				"example": {
					"success": true,
					"msg": "Doctor delete successfully",
					"data": []
				}
			},
			"socialloginrequest": {
				"title": "socialloginrequest",
				"required": ["token", "type"],
				"type": "object",
				"properties": {
					"token": {
						"type": "string"
					},
					"type": {
						"type": "string"
					}
				},
				"example": {
					"token": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6ImNkNjcyYTI5OGRkNTk3NzYyMTJlNmE2ZDgzZmI4YTllNjBmZjNiZWYifQ.eyJpc3MiOiJodHRwczpcL1wvd3d3LmZhY2Vib29rLmNvbSIsImF1ZCI6IjEwOTYwMTA0MzQzMzgxOTIiLCJzdWIiOiIxNDUxMTA5MTg1MzAxNTU0IiwiaWF0IjoxNjU4MzIyMTI3LCJleHAiOjE2NTgzMjU3MjcsImp0aSI6IkRJMFguZDlmMWU3ODZjOTc5OTNkNTk1ODMxZTc1ZTAyNzc0YmNmZDQyODVjMDE5NzFkOGQ4M2NmNWE1M2UzNTJkZTEzMiIsIm5vbmNlIjoibXlfbm9uY2UiLCJlbWFpbCI6InBpcGFsaXlhc2Fua2V0MVx1MDA0MGdtYWlsLmNvbSIsImdpdmVuX25hbWUiOiJTYW5rZXQiLCJmYW1pbHlfbmFtZSI6IlBpcGFsaXlhIiwibmFtZSI6IlNhbmtldCBQaXBhbGl5YSIsInBpY3R1cmUiOiJodHRwczpcL1wvcGxhdGZvcm0tbG9va2FzaWRlLmZic2J4LmNvbVwvcGxhdGZvcm1cL3Byb2ZpbGVwaWNcLz9hc2lkPTE0NTExMDkxODUzMDE1NTQmaGVpZ2h0PTEwMCZ3aWR0aD0xMDAmZXh0PTE2NjA5MTQxMjgmaGFzaD1BZVNVUmR5QVYzX3NaLVkwRDBVIn0.fzXmN3ivyVpq3cWjPV94lgk3iimF6dsl4PnhfvGj2WntwXG8VUJ6o_VEM_Tau_Un6Un_JUl1TnKMM-4ljtCNlEn4YXfGpSHXwatZ22X7UyRjxTTBZezloUw_sFrolSWz_ys9NsKWfGvO8WBLbekppE6bJNUzVe-jCrkLmY7fOVc2IguVg36VWcyIsZAvdeGzsIC5rm3D9jRg8h2HJuTpH5NB2X58Vq8enMv9ZzwGKmRkkFmpUDEyMXGMUZ6eH622-Ngn2hfHIQJIaJFMLEVNxpQMD7GGKNyKSRpuqHK-E-T95Ug-KG1wha6hwi-G_t8So7T2Vqpl8XOCcpJMVRd_Jw",
					"type": "CUSTOMER"
				}
			},
...
			"m2007": {
				"title": "m2007",
				"required": ["success", "message", "data"],
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"description": ""
					}
				},
				"example": {
					"success": true,
					"message": "Device created successfully",
					"data": []
				}
			},
...
			"UpdatedeviceRequest": {
				"title": "UpdatedeviceRequest",
				"required": ["deviceName", "deviceType", "deviceModal", "treatment"],
				"type": "object",
				"properties": {
					"deviceName": {
						"type": "string"
					},
					"deviceType": {
						"type": "string"
					},
					"deviceModal": {
						"type": "string"
					},
					"treatment": {
						"type": "string"
					}
				},
				"example": {
					"deviceName": "fddddff",
					"deviceType": "aaa",
					"deviceModal": "tm005a",
					"treatment": "aaaa"
				}
			},
...
		},
		"securitySchemes": {
			"bearer": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"security": [],
...
}